requirements.txt
django==4.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
drf-spectacular==0.27.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pillow==11.0.0
flake8==7.1.1
